pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = ""
authors = [
    {name = "uncleben006",email = "uncleben006@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain-community (>=0.3.19,<0.4.0)",
    "langchain-cli (>=0.0.35,<0.0.36)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "qdrant-client (>=1.13.2,<2.0.0)",
    "pypdf (>=5.3.1,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
